完成了歌曲管理与日志管理

+++ src/main/java/com/example/webmusic/controller/song/in/InApiGetPageSong.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 @Builder
 public class InApiGetPageSong {
-    private int currentPage;
-    private int pageSize;
+    private long currentPage;
+    private long pageSize;
 }
